+++ packages/cargo/jns42-core/src/lib.rs
@@ -1,5 +1,5 @@
 mod ffi;
 pub mod models;
-pub mod names;
+pub mod naming;
 pub mod schema_transforms;
 pub mod utils;
