+++ packages/cargo/jns42-core/src/naming/names.rs
