+++ packages/cargo/jns42-core/src/naming/names_builder.rs
