+++ packages/cargo/jns42-core/src/naming/part.rs
