+++ packages/cargo/jns42-core/src/naming/sentence.rs
