+++ packages/cargo/jns42-core/src/naming/word.rs
